Add files via upload
This project is about handling and managing the Controller of Examinations of a college. All the provisions like student detail management, results, hall tickets, time tables, forms such as transport, no-due, on duty forms, leave forms are also made available in our project. The students can get all their requirements from this project. It is easier for the staff members to add, delete and modify the student records, and calculate the results , GPA and CGPA too.